LeagueAPI and LeagueListAPI tests; added error handling and exceptions to the league model and APIs

+++ src/app/league/models.py
@@ -1,3 +1,4 @@
+from app.league.exceptions import InvalidRatingScheme, LeagueNotFound
 import tinyid
 from google.appengine.ext import ndb
 from app.models import BaseModel
@@ -33,10 +34,10 @@ class LeagueModel(BaseModel):
 
 
 def create_league(user, name, rating_scheme, description=None):
+    if rating_scheme not in LeagueModel.rating_scheme._choices:
+        raise ValueError("Invalid rating scheme %s" % rating_scheme)
     league_id = LeagueModel.generate_id()
     key = LeagueModel.build_key(league_id, user.key)
-    if rating_scheme not in LeagueModel.rating_scheme._choices:
-        raise ValueError
     new_league = LeagueModel(key=key, league_id=league_id, name=name, rating_scheme=rating_scheme,
                              description=description)
     new_league.put()
@@ -50,8 +51,11 @@ def update_league(user, league_id, name, rating_scheme, description=None):
 
     key = LeagueModel.build_key(league_id, user.key)
     league = key.get()
+    
     if not league:
-        raise ValueError("There is no league for league_id %s" % league_id)
+        raise LeagueNotFound("There is no league for league_id %s" % league_id)
+    if rating_scheme not in LeagueModel.rating_scheme._choices:
+        raise InvalidRatingScheme("Invalid rating scheme %s" % rating_scheme)
 
     league.name = name
     league.rating_scheme = rating_scheme
@@ -63,4 +67,7 @@ def update_league(user, league_id, name, rating_scheme, description=None):
 
 def delete_league(user, league_id):
     key = LeagueModel.build_key(league_id, user.key)
-    return key.delete()
+    if key.get() is None:
+        raise LeagueNotFound
+    else:
+        return key.delete()
